implemented Inventory Domain, Repository, Aplication cantracts and concretes.

+++ LampShade/InventoryManagement.Application.Contracts/InventoryAgg/CreateInventory.cs
@@ -3,6 +3,6 @@
     public class CreateInventory
     {
         public long ProductId { get; set; }
-        public long UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
     }
 }
